# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Alberta Government Internal API",
    "description": "Internal-facing API for Alberta Government enterprise template.\n\nSupports two authentication models:\n- **Staff auth** (session-based): MS Entra ID or Mock driver, session cookie\n- **Service-to-service auth** (Bearer token): Azure AD Client Credentials flow\n\n### CSRF Protection\nState-changing requests (POST/PUT/PATCH/DELETE) require a CSRF token.\n1. Call `GET /api/v1/csrf-token` to obtain a token\n2. Include it as the `X-CSRF-Token` header on subsequent requests",
    "version": "1.0.0",
    "contact": {
      "name": "Alberta Government"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie set after login via /api/v1/auth/login flow. Log in through the browser first; the cookie is sent automatically."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF token from GET /api/v1/csrf-token. Required for POST/PUT/PATCH/DELETE requests."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Azure AD access token obtained via OAuth 2.0 Client Credentials flow."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "UNAUTHORIZED" },
              "message": { "type": "string", "example": "Authentication required" },
              "details": { "type": "string" }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["success", "error"]
      },
...
    }
  },
  "paths": {
...
    "/api/v1/auth/callback": {
      "get": {
        "tags": ["Auth"],
        "summary": "Authentication callback",
        "description": "Handles the callback from the identity provider. Sets session cookie and redirects to frontend. Rate limited.",
        "operationId": "authCallback",
        "security": [],
        "parameters": [
          {
            "name": "mockUserId",
            "in": "query",
            "description": "Mock user ID (mock driver only). Available: mock-user-1 (developer), mock-user-2 (admin), mock-user-3 (user)",
            "schema": { "type": "string", "example": "mock-user-1" }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to frontend after successful authentication",
            "headers": {
              "Location": {
                "description": "Frontend profile page URL",
                "schema": { "type": "string", "example": "http://localhost:5173/profile" }
              },
              "Set-Cookie": {
                "description": "Session cookie",
                "schema": { "type": "string" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "List users (example)",
        "description": "Example role-protected endpoint. Requires the `admin` role. Replace with your admin logic.",
        "operationId": "getAdminUsers",
        "security": [
          { "sessionAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Admin endpoint response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions (requires admin role)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "success": false,
                  "error": { "code": "FORBIDDEN", "message": "Insufficient permissions", "details": "Required role(s): admin" }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/public/info": {
      "get": {
        "tags": ["Public API"],
        "summary": "Public API metadata",
        "description": "Returns API metadata and the authenticated service client's identity.",
        "operationId": "getPublicInfo",
        "security": [
          { "bearerAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Public API info with client identity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "name": { "type": "string", "example": "Alberta Government Internal API" },
                        "version": { "type": "string", "example": "v1" },
                        "client": {
                          "type": "object",
                          "properties": {
                            "clientId": { "type": "string", "example": "abc12345-def6-7890-ghij-klmnopqrstuv" },
                            "roles": {
                              "type": "array",
                              "items": { "type": "string" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "503": {
            "description": "Service authentication not configured",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  }
}